+++ recipes/cpp.rb
@@ -1,0 +1,14 @@
+if node['protobuf']['install_type'] == "archive"
+  include_recipe "protobuf::archive"
+else
+  protobuf_recipe = value_for_platform(
+    %w{centos fedora redhat} => {
+      "default" => "package_cpp"
+    },
+    %w{ubuntu} => {
+      "default" => "package_cpp"
+    }
+  )
+
+  include_recipe "protobuf::#{protobuf_recipe}"
+end