+++ recipes/package.rb
@@ -1,3 +1,0 @@
-include_recipe "protobuf::package_cpp"
-include_recipe "protobuf::package_java"
-include_recipe "protobuf::package_python"